[refactor] move java package to namespace java
