+++ jvm-packages/checkstyle-suppressions.xml
@@ -29,5 +29,5 @@
 
 <suppressions>
 <suppress checks=".*"
-	files="xgboost4j/src/main/java/ml/dmlc/xgboost4j/XgboostJNI.java"/>
+	files="xgboost4j/src/main/java/ml/dmlc/xgboost4j/java/XgboostJNI.java"/>
 </suppressions>
