+++ jvm-packages/xgboost4j-demo/src/main/java/ml/dmlc/xgboost4j/java/demo/GeneralizedLinearModel.java
@@ -13,54 +13,58 @@
  See the License for the specific language governing permissions and
  limitations under the License.
  */
-package ml.dmlc.xgboost4j.demo;
+package ml.dmlc.xgboost4j.java.demo;
 
-import java.util.Arrays;
 import java.util.HashMap;
 
-import ml.dmlc.xgboost4j.Booster;
-import ml.dmlc.xgboost4j.DMatrix;
-import ml.dmlc.xgboost4j.XGBoost;
-import ml.dmlc.xgboost4j.XGBoostError;
+import ml.dmlc.xgboost4j.java.Booster;
+import ml.dmlc.xgboost4j.java.DMatrix;
+import ml.dmlc.xgboost4j.java.XGBoost;
+import ml.dmlc.xgboost4j.java.XGBoostError;
+import ml.dmlc.xgboost4j.java.demo.util.CustomEval;
 
 /**
- * predict leaf indices
+ * this is an example of fit generalized linear model in xgboost
+ * basically, we are using linear model, instead of tree for our boosters
  *
  * @author hzx
  */
-public class PredictLeafIndices {
+public class GeneralizedLinearModel {
   public static void main(String[] args) throws XGBoostError {
     // load file from text file, also binary buffer generated by xgboost4j
     DMatrix trainMat = new DMatrix("../../demo/data/agaricus.txt.train");
     DMatrix testMat = new DMatrix("../../demo/data/agaricus.txt.test");
 
     //specify parameters
+    //change booster to gblinear, so that we are fitting a linear model
+    // alpha is the L1 regularizer
+    //lambda is the L2 regularizer
+    //you can also set lambda_bias which is L2 regularizer on the bias term
     HashMap<String, Object> params = new HashMap<String, Object>();
-    params.put("eta", 1.0);
-    params.put("max_depth", 2);
+    params.put("alpha", 0.0001);
     params.put("silent", 1);
     params.put("objective", "binary:logistic");
+    params.put("booster", "gblinear");
+
+    //normally, you do not need to set eta (step_size)
+    //XGBoost uses a parallel coordinate descent algorithm (shotgun),
+    //there could be affection on convergence with parallelization on certain cases
+    //setting eta to be smaller value, e.g 0.5 can make the optimization more stable
+    //param.put("eta", "0.5");
+
 
     //specify watchList
     HashMap<String, DMatrix> watches = new HashMap<String, DMatrix>();
     watches.put("train", trainMat);
     watches.put("test", testMat);
 
-
     //train a booster
-    int round = 3;
+    int round = 4;
     Booster booster = XGBoost.train(params, trainMat, round, watches, null, null);
 
-    //predict using first 2 tree
-    float[][] leafindex = booster.predict(testMat, 2, true);
-    for (float[] leafs : leafindex) {
-      System.out.println(Arrays.toString(leafs));
-    }
+    float[][] predicts = booster.predict(testMat);
 
-    //predict all trees
-    leafindex = booster.predict(testMat, 0, true);
-    for (float[] leafs : leafindex) {
-      System.out.println(Arrays.toString(leafs));
-    }
+    CustomEval eval = new CustomEval();
+    System.out.println("error=" + eval.eval(predicts, testMat));
   }
 }
