+++ jvm-packages/xgboost4j-demo/src/main/java/ml/dmlc/xgboost4j/java/demo/PredictLeafIndices.java
@@ -13,27 +13,26 @@
  See the License for the specific language governing permissions and
  limitations under the License.
  */
-package ml.dmlc.xgboost4j.demo;
+package ml.dmlc.xgboost4j.java.demo;
 
+import java.util.Arrays;
 import java.util.HashMap;
 
-import ml.dmlc.xgboost4j.Booster;
-import ml.dmlc.xgboost4j.DMatrix;
-import ml.dmlc.xgboost4j.XGBoost;
-import ml.dmlc.xgboost4j.XGBoostError;
+import ml.dmlc.xgboost4j.java.Booster;
+import ml.dmlc.xgboost4j.java.DMatrix;
+import ml.dmlc.xgboost4j.java.XGBoost;
+import ml.dmlc.xgboost4j.java.XGBoostError;
 
 /**
- * simple example for using external memory version
+ * predict leaf indices
  *
  * @author hzx
  */
-public class ExternalMemory {
+public class PredictLeafIndices {
   public static void main(String[] args) throws XGBoostError {
-    //this is the only difference, add a # followed by a cache prefix name
-    //several cache file with the prefix will be generated
-    //currently only support convert from libsvm file
-    DMatrix trainMat = new DMatrix("../../demo/data/agaricus.txt.train#dtrain.cache");
-    DMatrix testMat = new DMatrix("../../demo/data/agaricus.txt.test#dtest.cache");
+    // load file from text file, also binary buffer generated by xgboost4j
+    DMatrix trainMat = new DMatrix("../../demo/data/agaricus.txt.train");
+    DMatrix testMat = new DMatrix("../../demo/data/agaricus.txt.test");
 
     //specify parameters
     HashMap<String, Object> params = new HashMap<String, Object>();
@@ -42,20 +41,26 @@ public class ExternalMemory {
     params.put("silent", 1);
     params.put("objective", "binary:logistic");
 
-    //performance notice: set nthread to be the number of your real cpu
-    //some cpu offer two threads per core, for example, a 4 core cpu with 8 threads, in such case
-    // set nthread=4
-    //param.put("nthread", num_real_cpu);
-
     //specify watchList
     HashMap<String, DMatrix> watches = new HashMap<String, DMatrix>();
     watches.put("train", trainMat);
     watches.put("test", testMat);
 
-    //set round
-    int round = 2;
 
-    //train a boost model
+    //train a booster
+    int round = 3;
     Booster booster = XGBoost.train(params, trainMat, round, watches, null, null);
+
+    //predict using first 2 tree
+    float[][] leafindex = booster.predict(testMat, 2, true);
+    for (float[] leafs : leafindex) {
+      System.out.println(Arrays.toString(leafs));
+    }
+
+    //predict all trees
+    leafindex = booster.predict(testMat, 0, true);
+    for (float[] leafs : leafindex) {
+      System.out.println(Arrays.toString(leafs));
+    }
   }
 }
