+++ jvm-packages/xgboost4j-flink/src/main/scala/ml/dmlc/xgboost4j/flink/Test.scala
@@ -16,8 +16,7 @@
 
 package ml.dmlc.xgboost4j.flink
 
-import ml.dmlc.xgboost4j.Rabit
-import ml.dmlc.xgboost4j.RabitTracker
+import ml.dmlc.xgboost4j.java.{RabitTracker, Rabit}
 import ml.dmlc.xgboost4j.scala.Booster
 import ml.dmlc.xgboost4j.scala.DMatrix
 import ml.dmlc.xgboost4j.scala.XGBoost
