+++ jvm-packages/xgboost4j/src/main/java/ml/dmlc/xgboost4j/java/XgboostJNI.java
@@ -13,7 +13,7 @@
  See the License for the specific language governing permissions and
  limitations under the License.
  */
-package ml.dmlc.xgboost4j;
+package ml.dmlc.xgboost4j.java;
 
 
 /**
