+++ jvm-packages/xgboost4j/src/main/scala/ml/dmlc/xgboost4j/scala/Booster.scala
@@ -18,10 +18,8 @@ package ml.dmlc.xgboost4j.scala
 
 import java.io.IOException
 
+import ml.dmlc.xgboost4j.java.XGBoostError
 import scala.collection.mutable
-
-import ml.dmlc.xgboost4j.XGBoostError
-
 
 trait Booster extends Serializable {
