+++ jvm-packages/xgboost4j/src/main/scala/ml/dmlc/xgboost4j/scala/XGBoost.scala
@@ -18,8 +18,6 @@ package ml.dmlc.xgboost4j.scala
 
 import scala.collection.JavaConverters._
 
-import ml.dmlc.xgboost4j.{XGBoost => JXGBoost}
-
 object XGBoost {
 
   def train(
