+++ jvm-packages/xgboost4j/src/test/scala/ml/dmlc/xgboost4j/scala/DMatrixSuite.scala
@@ -20,7 +20,6 @@ import java.util.Arrays
 
 import scala.util.Random
 
-import ml.dmlc.xgboost4j.{DMatrix => JDMatrix}
 import org.scalatest.FunSuite
 
 class DMatrixSuite extends FunSuite {
